chore: restructure #11
