+++ src/features/auth/views/LoginView.vue
