+++ src/features/game/multiplayer/room/components/settings/ModePicker.vue
