+++ src/features/game/single/main/views/PhotoModeView.vue
@@ -156,7 +156,7 @@ import { ref, computed } from 'vue';
 import { useRouter } from 'vue-router';
 import AppLogo from "@/core/components/BaseAppLogo.vue";
 import useGame from '@/composables/useGame';
-import GameModeCard from "@/components/common/ui/GameModeCard.vue";
+import GameModeCard from "@/features/game/shared/components/Common/GameModeCard.vue";
 import useAuth from '@/composables/useAuth';
 
 // 라우터 및 컴포지션 API 사용
