+++ src/features/game/single/photo/components/Hint/HintDisplay.vue
