+++ src/features/game/single/photo/components/Photo/PhotoGrid.vue
