+++ src/features/game/single/photo/components/Result/GameResult.vue
