+++ src/features/game/single/photo/views/PracticeView.vue
