+++ src/features/game/single/roadview/views/RankView.vue
@@ -101,11 +101,11 @@
 </template>
 
 <script>
-import RoadViewGame from "./components/gameplay/RoadViewGame.vue";
+import RoadViewGame from "../components/gameplay/RoadViewGame.vue";
 import PhoneFrame from "@/components/game/phone/PhoneFrame.vue";
-import CountdownOverlay from "@/components/game/countdown/CountdownOverlay.vue";
-import IntroOverlay from "@/components/game/intro/IntroOverlay.vue";
-import ResultOverlay from "./components/result/ResultOverlay.vue";
+import CountdownOverlay from "@/features/game/shared/components/Common/CountdownOverlay.vue";
+import IntroOverlay from "@/features/game/shared/components/Common/IntroOverlay.vue";
+import ResultOverlay from "../components/Result/ResultOverlay.vue";
 
 export default {
   name: "RoadViewRank",
