+++ src/features/user/components/AccountSettings.vue
@@ -253,7 +253,7 @@
 </template>
 
 <script>
-import NavigationBar from '@/components/common/NavigationBar.vue';
+import NavigationBar from '@/core/components/NavigationBar.vue';
 
 export default {
   name: 'AccountSettings',
