+++ src/features/user/components/FriendsList.vue
@@ -222,7 +222,7 @@
 </template>
 
 <script>
-import NavigationBar from '@/components/common/NavigationBar.vue';
+import NavigationBar from '@/core/components/NavigationBar.vue';
 
 export default {
   name: 'FriendsList',
