+++ src/features/user/view/ProfileView.vue
@@ -101,7 +101,7 @@
 </template>
 
 <script>
-import NavigationBar from '@/components/common/NavigationBar.vue';
+import NavigationBar from '@/core/components/NavigationBar.vue';
 
 export default {
   name: 'MyProfile',
