+++ src/views/Game/MultiplayerMode/components/gameplay/TestVoting.vue
@@ -50,7 +50,7 @@
 <script setup>
 import { computed } from 'vue';
 import TeamPlayersList from 'src/views/Game/MultiplayerMode/components/gameplay/roadview/components/TeamPlayersList.vue'
-import BasePlayerCard from '@/components/common/ui/game/BasePlayerCard.vue';
+import BasePlayerCard from '@/features/game/shared/components/Player/Card.vue';
 import gameStore from '@/store/gameStore';
 
 // Load test data when component mounts
